Add FURscript HELLO WORLD
That should work. I was unable to test due to no compiler for this "language"
actually existing. Maybe that's a good thing. I'm not really sure why it even
exists in the first place. There's no flow control, the specification is
laughable... Regardless, it's a thing. And now someone wrote a hello world (or
rather A HELLO WORLD) script in it.

Also, reference to issue #1